Fix conflict code

+++ banking/src/main/java/com/project88/banking/controller/UserController.java
@@ -107,9 +107,9 @@ public class UserController {
         depositService.createDeposit(form, userId);
 
     }
-    
+
     @GetMapping("/{id}/balance")
-    public ResponseEntity<Integer> getUserBalance(@PathVariable(name = "id") short id) {
+    public ResponseEntity<Integer> getUserBalance(@PathVariable(name = "id") long id) {
         User user = userService.findUserById(id);
         return new ResponseEntity<>(user.getBalance(), HttpStatus.OK);
     }
